claude-telegram-bridge-package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-telegram-bridge"
version = "1.0.0"
description = "A modular bridge system for connecting Claude Code with Telegram for monitoring and control"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Claude-Ops Team", email = "noreply@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent", 
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
keywords = ["claude", "telegram", "bot", "monitoring", "bridge", "ai"]
requires-python = ">=3.8"
dependencies = [
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0", 
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/kyuwon-shim-ARL/claude-ops"
Documentation = "https://github.com/kyuwon-shim-ARL/claude-ops/blob/main/claude_bridge/README.md"
Repository = "https://github.com/kyuwon-shim-ARL/claude-ops"
Issues = "https://github.com/kyuwon-shim-ARL/claude-ops/issues"

[project.scripts]
claude-bridge = "claude_telegram_bridge.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["claude_telegram_bridge"]

[tool.hatch.build.targets.sdist]
include = [
    "claude_telegram_bridge/",
    "README.md",
    "LICENSE",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true